requirements.txt
cryptography
pycryptodome
rsa
base64
secrets
hashlib
Crypto.Cipher